+++ scripts/azureml-assets/CHANGELOG.md
@@ -2,6 +2,14 @@
 ### 🚀 New Features
 ### 🐛 Bugs Fixed
 
+## 1.16.91 (2025-09-19)
+### 🚀 New Features
+- [#4448](https://github.com/Azure/azureml-assets/pull/4448) Add EVALUATOR to AssetType
+
+## 1.16.90 (2025-09-12)
+### 🚀 New Features
+- [#4434](https://github.com/Azure/azureml-assets/pull/4434) Support system_metadata by using the SDK for model create and update
+
 ## 1.16.89 (2025-08-18)
 ### 🐛 Bugs Fixed
 - [#4391](https://github.com/Azure/azureml-assets/pull/4391) Pass in correct uri to LocalAssetPath from ModelConfig